Cleanup the ochttp package
Several reorganization related and readability improvements.
